paddleformers/peft/lora/utils.py
# Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

from contextlib import nullcontext

from paddle.distributed.fleet.meta_parallel import get_rng_state_tracker


def rng_ctx(is_mp: bool, in_dynamic_mode: bool):
    return get_rng_state_tracker().rng_state() if (is_mp and in_dynamic_mode) else nullcontext()
