requirements-dev.txt
paddleocr
simsimd==1.1.2
pre-commit
pytest==8.1.1
parameterized
pytest-cov
regex
pytest-xdist
pytest-timeout
emoji
ftfy
yacs
unidecode
soundfile
librosa
numpy
rouge
tiktoken
visualdl
wandb
swanlab
tensorboard
tensorboardX
modelscope
hyperopt
h5py
deploy
ray
loguru
data
wget
huggingface_hub>=0.19.2
tiktoken
tokenizers<=0.20.3; python_version<="3.8"
tokenizers>=0.21,<0.22; python_version>"3.8"